+++ ckanext/harvest/utils.py
@@ -232,6 +232,18 @@ def clear_harvest_source_history(source_id, keep_current):
             len(cleared_sources_dicts))
 
 
+def purge_harvest_source(source_id_or_name):
+    context = {
+        "model": model,
+        "session": model.Session,
+        "user": _admin_user()["name"],
+    }
+    source = tk.get_action("harvest_source_show")(context, {
+        "id": source_id_or_name
+    })
+    tk.get_action("purge_harvest_source")(context, {"id": source["id"]})
+
+
 def abort_failed_jobs(job_life_span, include, exclude):
     context = {
         "model": model,
@@ -694,6 +706,22 @@ def clear_view(id):
         h.url_for('{0}_admin'.format(DATASET_TYPE_NAME), id=id))
 
 
+def purge_view(id):
+    try:
+        context = {'model': model, 'user': tk.c.user, 'session': model.Session}
+        tk.get_action('purge_harvest_source')(context, {'id': id})
+        h.flash_success(_('Harvest source purged'))
+    except tk.ObjectNotFound:
+        return tk.abort(404, _('Harvest source not found'))
+    except tk.NotAuthorized:
+        return tk.abort(401, _not_auth_message())
+    except Exception as e:
+        msg = 'An error occurred: [%s]' % str(e)
+        h.flash_error(msg)
+
+    return h.redirect_to('/harvest')
+
+
 def delete_view(id):
     try:
         context = {'model': model, 'user': tk.c.user}
@@ -703,8 +731,17 @@ def delete_view(id):
                                                             u'true',
                                                             u'1',
                                                         )
+        context['purge_resource'] = tk.request.params.get('purge',
+                                                          '').lower() in (
+                                                            u'true',
+                                                            u'1',
+                                                        )
 
         tk.get_action('harvest_source_delete')(context, {'id': id})
+
+        if context['purge_resource']:
+            h.flash_success(_('Harvesting source successfully purged'))
+            return h.redirect_to('/harvest')
 
         if context['clear_source']:
             h.flash_success(_('Harvesting source successfully cleared'))